Move errors badge code to development folder
