+++ src/background.ts
@@ -18,6 +18,8 @@
 // extensionContext needs to be imported before webpack-target-webextension to
 // ensure the webpack path is correct
 import "@/extensionContext";
+import "@/development/autoreload";
+import "@/development/errorsBadge";
 
 // Required for MV3; Service Workers don't have XMLHttpRequest
 import "@/background/axiosFetch";
@@ -25,7 +27,6 @@ import "@/background/axiosFetch";
 import "webext-dynamic-content-scripts";
 
 import "@/background/messenger/registration";
-import "@/development/autoreload";
 import "@/messaging/external";
 import "@/background/locator";
 import "@/background/contextMenus";
